change test description and refactor the ui structure
