+++ packages/core/content-releases/admin/src/pages/ReleasesPage/Releases.tsx
@@ -5,7 +5,7 @@ import { useNotification } from '@strapi/helper-plugin';
 import { Plus } from '@strapi/icons';
 import { useIntl } from 'react-intl';
 
-import { AddReleaseDialog } from '../components/AddReleaseDialog';
+import { AddReleaseDialog } from '../../components/AddReleaseDialog';
 
 const ReleasesPage = () => {
   const [addReleaseDialogIsShown, setAddReleaseDialogIsShown] = React.useState(false);
